Actually, make fix a dword so no padding is necessary.

+++ src/module_info.h
@@ -71,7 +71,7 @@ struct DeviceStatus {
 };
 
 struct DeviceLocation {
-    uint8_t fix;
+    uint32_t fix;
     uint32_t time;
     float coordinates[MaximumCoordinates];
